Refactor: Match, Patient, Caregiver 패키지 조정
