+++ src/main/java/com/patientpal/backend/caregiver/repository/CaregiverRepository.java
@@ -1,5 +1,7 @@
-package com.patientpal.backend.member.domain;
+package com.patientpal.backend.caregiver.repository;
 
+import com.patientpal.backend.caregiver.domain.Caregiver;
+import com.patientpal.backend.member.domain.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
