+++ src/main/java/com/patientpal/backend/matching/dto/response/MatchResponse.java
@@ -4,9 +4,8 @@ import com.patientpal.backend.matching.domain.FirstRequest;
 import com.patientpal.backend.matching.domain.Match;
 import com.patientpal.backend.matching.domain.MatchStatus;
 import com.patientpal.backend.matching.domain.ReadStatus;
-import com.patientpal.backend.member.domain.Caregiver;
-import com.patientpal.backend.member.domain.Member;
-import com.patientpal.backend.member.domain.Patient;
+import com.patientpal.backend.caregiver.domain.Caregiver;
+import com.patientpal.backend.patient.domain.Patient;
 import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
