+++ src/main/java/com/patientpal/backend/patient/domain/Patient.java
@@ -1,6 +1,9 @@
-package com.patientpal.backend.member.domain;
+package com.patientpal.backend.patient.domain;
 
 import com.patientpal.backend.matching.domain.Match;
+import com.patientpal.backend.member.domain.Address;
+import com.patientpal.backend.member.domain.Member;
+import com.patientpal.backend.patient.dto.PatientProfileRequest;
 import jakarta.persistence.*;
 import lombok.*;
 
@@ -54,5 +57,15 @@ public class Patient {
         this.nokContact = nokContact;
         this.patientSignificant = patientSignificant;
         this.careRequirements = careRequirements;
+    }
+
+    public void registerDetailProfile(PatientProfileRequest patientProfileRequest) {
+        this.name = patientProfileRequest.getName();
+        this.residentRegistrationNumber = patientProfileRequest.getResidentRegistrationNumber();
+        this.address = patientProfileRequest.getAddress();
+        this.nokName = patientProfileRequest.getNokName();
+        this.nokContact = patientProfileRequest.getNokContact();
+        this.patientSignificant = patientProfileRequest.getPatientSignificant();
+        this.careRequirements = patientProfileRequest.getCareRequirements();
     }
 }
